LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_hit_test_scroll-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "wheel".

Harness Error (FAIL), message = Unhandled rejection: Unknown source type "wheel".

TIMEOUT Wheel-scroll over pointer-events: none scroller skips that scroller Test timed out

Harness Error (FAIL), message = Unhandled rejection: Unknown source type "wheel".

TIMEOUT Wheel-scroll over pointer-events: none scroller skips that scroller Test timed out
